HTBHF-2569 Move v1 eligibility specific classes into a v1 package.
